Git merge master into 60-ne-seed-data
I ended up doing this twice.  I was trying to ensure that the merge went
smoothly be first merging (locally) my branch into master.

It was not the smooth process I had hoped for, and I basically ended up
resolving the (slightly different) merge conflicts twice.  Not the most
fun I've had on a Sunday night.